refactor(connect): move ethereum method definitions to corresponding folder
