+++ packages/connect-common/src/types/api/ethereum/ethereumGetAddress.ts
@@ -1,4 +1,4 @@
-import type { Address, BundledParams, GetAddress, Params, Response } from '../params';
+import type { Address, BundledParams, GetAddress, Params, Response } from '../../params';
 
 export declare function ethereumGetAddress(params: Params<GetAddress>): Response<Address>;
 export declare function ethereumGetAddress(params: BundledParams<GetAddress>): Response<Address[]>;
