+++ packages/connect-common/src/types/api/ethereum/ethereumSignTransaction.ts
@@ -1,0 +1,6 @@
+import type { EthereumSignTransaction, EthereumSignedTx } from './common';
+import type { Params, Response } from '../../params';
+
+export declare function ethereumSignTransaction(
+    params: Params<EthereumSignTransaction>,
+): Response<EthereumSignedTx>;
