+++ packages/connect-common/src/types/api/ethereum/ethereumSignTypedData.ts
@@ -1,11 +1,11 @@
 import type { MessagesSchema as PROTO } from '@trezor/protobuf';
 
-import type { Params, Response } from '../params';
 import type {
     EthereumSignTypedData,
     EthereumSignTypedDataTypes,
     EthereumSignTypedHash,
-} from './ethereum';
+} from './common';
+import type { Params, Response } from '../../params';
 
 export declare function ethereumSignTypedData<T extends EthereumSignTypedDataTypes>(
     params: Params<EthereumSignTypedData<T>>,
