+++ packages/connect-common/src/types/api/ethereum/ethereumVerifyMessage.ts
@@ -1,7 +1,7 @@
 import type { MessagesSchema as PROTO } from '@trezor/protobuf';
 
-import type { Params, Response } from '../params';
-import type { EthereumVerifyMessage } from './ethereum';
+import type { EthereumVerifyMessage } from './common';
+import type { Params, Response } from '../../params';
 
 export declare function ethereumVerifyMessage(
     params: Params<EthereumVerifyMessage>,
