--- packages/connect-common/src/types/api/ethereumSignTransaction.ts
@@ -1,6 +1,0 @@
-import type { Params, Response } from '../params';
-import type { EthereumSignTransaction, EthereumSignedTx } from './ethereum';
-
-export declare function ethereumSignTransaction(
-    params: Params<EthereumSignTransaction>,
-): Response<EthereumSignedTx>;
